Move Prop3DMixin to prop3d.py
